Add files via upload
India grows enough food to meet the needs of its entire population, yet is unable  to feed millions of them.
The key reason is the poor maintenance and condition of cold storages.
What if, we upgrade the existing cold storage units technically ?
A system which is technically smart for both owner and user side.
What if, we create cold chain facilities to cut post-harvest losses?
A system which monitors the storage environment of cold storages and detect any  fault occurred.
So, out of many sectors such as hospitals, thermal power plants, etc. we have  chosen food processing sector especially cold storages.